Test that lens getters are scheduled on a specific queue
`TestScheduler` is known to be synchronous, which the previous version
of this test relies on to implicitly in order to assert that code is run
on a specific scheduler. It's not expected that `TestScheduler` would
ever become asynchronous, but it's valuable for a test about scheduling
code to make a stronger assertion about the code's execution context.

This change makes use of a private, serial test queue and
`DispatchQueue.setSpecific`/`DispatchQueue.getSpecific` to set up some
context that is completely under the control of the test, in order to
avoid false positives due to a data race between `TestScheduler` and the
test itself.